Workaround building with AVX
Cartographer only builds in 64bit platforms and without AVX support. The
reason is that they use Eigen, but they do not use
EIGEN_MAKE_ALIGNED_OPERATOR_NEW. Instead, they have decided not to
support AVX until that macro is not required anymore. That will happened
with C++17 and Eigen 3.4 (which has not been released yet).

This workaround has been suggested in:
https://github.com/cartographer-project/cartographer/issues/1001